Clarify the meaning of 'current' for consoles
This patch improves the documentation of the 'current' parameter of the
operations that retrieve the representation of virtual machine consoles.
It now explicitly explains that the 'address' and 'port' attributes
won't be populated unless 'current' is explicitly used with 'true' as
the value.

Change-Id: I0b0bd4b4507c69eb5d2906b8ef992c364e6355f8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 3116bfc6adf2b977b44a971ab30ba058b7aeb323)