Add Documentation for 'follow' input parameter
This patch adds to the in-memory represetation of the model documentation
of the 'follow' input parameter. This input paramter is relevant for the
HTTP GET requests, which are directed to get(), list() methods in the
API resources. Value of 'follow' indicates which inner links inside the
fetched entities should be 'followed', meaning their actual contents
fetched along with the enttiy.

Change-Id: I6e4ad28950a01735072513595decf0313bb8f039
Signed-off-by: Ori Liel <oliel@redhat.com>