Don't require deprecated affinity group attributes
The 'enforcing' and 'positive' attributes of affinity groups where
deprecteated in version 4.1 of the engine, but the documentation wasn't
updated, so they are still marked as required. This patch updates the
documentation and marks the attributes as deprecated.

Change-Id: I6cf7666e680468067bd98b1236fc52b213c38754
Bug-Url: https://bugzilla.redhat.com/1488729
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit d6c63a1381f04452ee8b8f7640656279147e351d)