Add external network provider configurations to host
If the automatic deployment of the external network provider is
supported, the networks of the referenced network provider are
available on the host.
Usually the same network provider is available on every host in
the cluster. This change enables to specify the network provider
on adding a host.

Even if currently only a single external network provider is
supported on a single host the external network provider is
encapsulated in a list, because this way a change in the REST-API
is avoided, if multiple external network providers are supported
on a host in future.

Change-Id: I89ef5ca9819183091ae0d376f839b5d813bab03b
Bug-Url: https://bugzilla.redhat.com/1495261
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit 75a5860943a03ece80c5b43ad995608d103d0444)