types: change S390 to S390X
The term S390 is historically used for the IBM Z architecture, with
S390X denoting the 64-bit variant. Marking the architecure explicitly
as S390X/64-bit aligns the type definitions with the other supported
architectures.
Besides changing the enum in Architecture, the description of
WatchdogModel.DIAG288 has been updated for consistency.

Change-Id: I75cedfb43156bf81d99ade109bc13d7d7ad84390
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>