Add mapping parameter to register operation
As part of the effort to support DR with oVirt
the "Resgister" operation is being added with a new mapping parameter
that describes the configuration of the registration.

The idea of supporting DR site to site in oVirt is to have 2 active
setups using storage replication between the primary setup and the
secondary setup.
Both setups will have active DCs, clusters, and hosts, although those
will not be indentical.
The user can define a mapping which will be used to recover its setup.

Each mapoing can be used to map any VM's/Template's attribute stored in the OVF
with its correlated entity.
For example, there could be a primary setup with a VM configured on cluster A.
We also keep an active secondary setup which only have cluster B.
Cluster B is compatible for that VM and in case of a DR scenario theoratically
the storage domain can be imported to the secondary setup and the use can
register the VM to cluster B.

In that case, we can automate the recovery process by defining a cluster mapping,
so once the entity will be registered its OVF will indicatre it belongs to
cluster A but the mapping which will be sent will indicate that cluster B should
be valid for every thing that is configured on cluster A.
The engine should do the switch and register the VM to cluster B in the secondary site.

Cluster mapping is just one example, there are different types of mappings:
  LUN mapping
  role mapping
  permissions mapping
  affinity group mapping
  ...

Each mapping will be used for its specific OVF's data once the register operation
will take place in the engine.

Change-Id: I68edbc9274a57de674c80d9461e6a0f02e979e32
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
(cherry picked from commit e4317f35ba192293d5b5603c64f61e846d0646e2)