Introduce the mix-in annotation
Currently when a method, parameter or attribute has to be repeated in
multiple places we have to actually repeat it, or else add it
automatically in all the relevant places of the model. A good example of
that is the recently introduced `follow` parameter, which should be
available in all the `Get` and `List` operations. In order to introduce
it we had two alternatives:

1. Explicitly add it to the model in all the relevant operatoins,
repeating it each time (approx 253 times with the current model).

2. Automatically add it when the model is processed by the metamodel
tools.

We decided to do the second, becaue the first was too much repetition.
But it has its drawbacks. The most important one is that the
documentation can't be managed like we manage every other piece of
documentatoin of the model, any update to the documentation needs a new
version of the metamodel tools. To address this drawback this patch
introduces the `Mixin` annotation. This annotation will be used for
interfaces that contain things that will be added to other interfaces
that extend it. For example, the `follow` parameter can now be described
in the model project with this mix-in:

  @Mixin
  public interface Follow {
    /**
     * Blah, blah, blah ...
     *
     * @author Jane Doe <jdoe@example.com>
     * @date 23 Oct 2017
     * @status added
     * @since 4.2
     */
     @In String follow();
  }

Then it can be used where needed, without repeating it:

  interface Get extends Follow {
    ...
  }

  interface List extends Follow {
    ...
  }

The down side of this is that it still requires the modification of the
253 places in the model where the `follow` parameter is supported, but I
think that the price is reasonable. The patch tha does that change is
here:

  Introduce the `Follow` mix-in
  https://gerrit.ovirt.org/83111

Change-Id: Idc79a3d4a659b9d6fd8cc07dd771b679eea084ec
Related-to: https://gerrit.ovirt.org/83111
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>