Uncatch WebFaultExceptions by Signature Invocation Code
Signature-detection and invocation auto-generated code
catches all exceptions and intereprets them as an
invocation failure. This is a bug. This patch makes
only exceptions which are related to the invocation
itself be interpreted in this way. WebFaultException
and WebApplicationExceptions, which may be thrown
for 'business' reasons, are propagated to the next catcher.

Bug-Url: https://bugzilla.redhat.com/1506284
Change-Id: Ic4b9d047e066c0479b261fde194da9fc57d62b78
Signed-off-by: Ori Liel <oliel@redhat.com>