Fix signature combining `or` and `mandatory`
Currently the generation of the signature detection code doesn't take
into account alternative mandatory parameters. For example, the
operation that exports a virtual machine to a storage domain includes
is described in th emodel like this:

  interface ToExportDoamin extends Export {
    @InputDetail
    default void inputDetail() {
      ...
      or(
        mandatory(storageDomain().id()),
        mandatory(storageDomain().name())
      );
    }
  }

(The meaning of that is that either the `id` or the `name` need to be
provided, otherwise that signature isn't matched.)

But the generated code only checks the first attribute inside the `or`,
the `id` in this particular case. As a result requests that only specify
the `name` aren't matched, which is wrong.

To address that this patch changes the code generator so that it takes
this particular case into account, and generates the right code.

Change-Id: I8ecb8aefe4d306c5b050d6d6d3ee0b5e01af0f89
Bug-Url: https://bugzilla.redhat.com/1523072
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>