restapi: Add lease parameter to preview snapshot
Add lease parameter to preview VM snapshot action.
The lease will specify the lease storage domain to
use in the preview.

The storage domains ids could be only the active snapshot
or the previewed snapshot lease storage domain ids.

In case the specify domain id does not match to the active
lease storage domain or to the previewed snapshot lease storage
domain validation error should occur.

If the lease attribute will be empty, the preview will not
contain any of the leases.

If the lease is missing the the previewed snapshot lease domain
will be used.

Change-Id: I3bc9d8d88aad3072ac9d0bc1141255accea7bb1b
Bug-Url: https://bugzilla.redhat.com/1532577
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>