build: checktyle for missing javadoc
As the API Model auto-generates its documentation from the javadoc,
methods and classes that are missing javadoc are an issue.

This patch adds a configuration of a checkstyle task to the build that
can be used to find such classes and methods.
It is currently NOT run by default, since it does not succeed (there
are methods missing their javadoc), but can be executed manually by
calling

$ mvn checkstyke:check

Once all the javadoc problems are resolved, this task can be added to
the check-patch.sh CI script to help ensure no new methods are added
without the proper javadoc.

Change-Id: I94b2f0b45228d60d934819332458e28679c7401f
Signed-off-by: Allon Mureinik <amureini@redhat.com>