Add note about update of external networks
Updating the name and the MTU of external networks updates
only oVirt's representation of the external network, the
update neither blocked nor it is propagated to the external
provider.

A note is added to make the user of the API aware of this.

Change-Id: Id678c7004b7ded230df368d8f3b0333721554122
Bug-Url: https://bugzilla.redhat.com/1451342
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit f1a14cb889ecace0a2138098f9ecbe35a9961f58)