api-model: poly dhcpv6 + autoconf iface config
A NIC ipv6 boot protocol configuration with both IPV6_AUTOCONF and
DHCPV6C set to 'yes' is not supported by engine, although it is a
legal configuration in linux OSs. Until now engine ignored the
IPV6_AUTOCONF=yes setting and set it to 'no'.

This patch introduces support for this 'poly' ipv6 boot protocol
option for REST API calls.

Bug-Url: https://bugzilla.redhat.com/1360839

Change-Id: I8b8780b3f9d77df7b06e1e48b95ae94568ee4375
Signed-off-by: Eitan Raviv <eraviv@redhat.com>