Compile sources prior to analyzing them
When building ovirt-engine-api-model, a maven plugin runs a tool from
ovirt-engine-api-metamodel project which analyzes model classes
and generates html and ascii-doc documentation from them.

Currently the analysis of the code is done prior to code compilation,
which may result in obscure error messages in case of complication errors.
This patch changes the order so that the code is compiled before analyzed.

Bug-Url: https://bugzilla.redhat.com/1533036
Change-Id: I3f0cff887ea6b751f40997b44fd38ff5b275d8c2
Signed-off-by: Ori_Liel <oliel@redhat.com>