Add Upgrade action to cluster
Add /api/clusters/1234/upgrade action that
sets the upgrade_running flag of the cluster to
true when upgrade_action 'start' is passed. If
the flag cannot be set because the flag has already
been set, then the action returns a response of false.

Add /api/clusters/1234/upgrade action that clears
the upgrade_running flag for the cluster when
upgrade_action 'finish' is passed in the action.

Change-Id: I09bcc0624795e11850f6a4a50f51ff4ceae7ba13
Bug-Url: https://bugzilla.redhat.com/1663626
Signed-off-by: Ravi Nori <rnori@redhat.com>