External disk support
Support creating a disk backed by external disk. This allows running a
VM from a read-only disk in any format supported by qemu. The external
disk can be a local file or a block device, or a network disk such as
NBD disk.

Interesting use cases for this configuration are:

- Importing VMs from other systems with minimal downtime.
- Preview a backup without copying the data to oVirt volume.
- With virt-v2v, import VMs from foreign systems with minimal downtime.
- Run transient VMs for testing on hosts with limited local storage.

Here is an example for creating a disk backed by NBD disk on another
host not managed by oVirt:

    disk=types.Disk(
        initial_size=6 * 1024**3,
        provisioned_size=50 * 1024**3,
        format=types.DiskFormat.COW,
        sparse=True,
        storage_domains=[
            types.StorageDomain(
                name='tmp-domain'
            )
        ],
        external_disk="nbd:example.com:9000"
    )

Starting a VM with this disk, the VM will write to the disk on the
"tmp-domain", and will read the image data from the NBD server at
"nbd:example.com:9000".

To import the external disk to oVirt storage domain, move the disk to
the destination storage domain:

    disk_service = disks_service.disk_service(disk.id)
    disks_service.move(
        storage_domain=types.StorageDomain(
            name="target-domain"
        )
    )

Change-Id: I1aa0af66bbdeb4a7e49d55465c96e806ebb23451
Signed-off-by: Nir Soffer <nsoffer@redhat.com>