Add new parameters to HostsService.List and VmService.Migrate
The new parameter in HostsService.List is 'checkVmsInAffinityClosure'.
When it is used together with 'migrationTargetOf', the result considers
VMs and all other VMs that are in positive affinity with them.

The new parameter in VmService.Migrate is 'migrateVmsInAffinityClosure'.
When it is set, the VM will be migrated together with all VMs that
are in positive enforcing affinity with it and run on the same host.

Change-Id: I76a3e4341f30170e7b1d6e2d09746558ee64df7e
Bug-Url: https://bugzilla.redhat.com/1651406
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>