api-model: echo external event to oVirt log on host
Write the event description to the oVirt log on a host when sending an
external event to engine.log.
If the event specifies a host the description is written to that host's
oVirt log only. Otherwise it is written to all hosts' oVirt logs.

This functionality facilitates correlating a host's log with engine.log

Change-Id: I6a557fe24ac7c12d0c6656c13b539da0f17eb54c
Signed-off-by: Eitan Raviv <eraviv@redhat.com>