restapi: introduce generic initialize
This patch will allow starting a VM with initialization.
The Initialization type will be determine using this flag
in the backend, by the operation system the VM is set.
If any of the initialization-types are explicitly set
(useCloudInit, useSysprep or useIgnition), they will be
prioritized and this flag will be ignored.

Change-Id: I0755c2787d58cd73b7a9c65de997b52e4b94f340
Bug-Url: https://bugzilla.redhat.com/1797659
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>