imageio: Remove ImageTransfer.signed_ticket
ImageTransfer.signed_ticket is unused since 4.2.0. In oVirt 4.4 engine
does not set this parameter.

Updated the wrong documentation suggesting to send this useless value in
the Authorization header when communicating with the proxy.

Also replace the complicated and wrong example code for transferring the
image with example of using imageio client library. This is not the
place to show how to use imageio API. For more details users should
check imageio documentation:
http://ovirt.github.io/ovirt-imageio/images.html

Change-Id: I1f82f5a4f61e14285673a9ce034c5494df6138f3
Bug-Url: https://bugzilla.redhat.com/1862722
Signed-off-by: Nir Soffer <nsoffer@redhat.com>