api-model: expose the VM OVF as OVA
This patch will allow the option to get the VM with OVF representation
as the OVA's OVF. This is useful in order to make external OVA creation.

The new boolean query parameter `ovf_as_ova` requires that
`all_content=True` will be set. Then, under the initialization, the OVF
representation will expose it as OVA's OVF.

Examples:
- GET /vms/{vm:id}?all_content=true&ovf_as_ova=true
- GET /vms?all_content=true&ovf_as_ova=true

Change-Id: I119f38a8dfb02265310209e273c7d7ad11dfe194
Bug-Url: https://bugzilla.redhat.com/1876234
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>