ImageTransfer: Support shallow download
Add "shallow" argument, allowing download of single image instead of the
entire image chain when using imageio client.download().

By default, transferring a disk or a disk snapshot transfers the entire
chain starting at the specified image. This is the recommended mode for
using incremental backup API.

Using shallow download, it is possible to download a single disk
snapshot, or the active disk snapshot of a disk (not exposed as a disk
snapshot yet). This will help users using snapshot based backup to
integrate with imageio client API now, and switch to incremental backup
API when the feature becomes more stable.

The typical use case is:

1. Create snapshot
2. Download the snapshot to qcow2 image
3. Delete an older snapshot

The result is a chain of images, one image per backup. Users can restore
the disk contents using any of the backups using imageio client.upload().

Change-Id: I12db565b7d236373ac0432ce80138bb6de5ba9d0
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>