incremental backup: added DiskBackupMode field to a Disk
The new DiskBackupMode field will be used during a backup operation,
DiskBackupMode can be:

    - FULL: When the disk is part of a full backup operation.
    - INCREMENTAL: When the disk part of an incremental backup operation.

The DiskBackupMode will indicate to the user the type of backup that
was taken for the disk in this backup operation, so the user will know
how to download this backup using imageio.

When the backup finalized, the DiskBackupMode will be set as 'null'.

Feature page: https://ovirt.org/develop/release-management/features/storage/incremental-backup.html

Change-Id: Ibaec8d2b1e322e8e191188e310d9c98b98014ea1
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>