storage: support finished tasks cleanup
Currently, the storage pool manager (SPM) fails to
switch to another host if the SPM has uncleared tasks.
Clearing all finished tasks enables the SPM switching.
Adding the REST API for that purpose.

For example, to clean all the finished tasks on a
data center with ID `123` send a request like this:

POST /ovirt-engine/api/datacenters/123/cleanfinishedtasks

With a request body like this:

<action/>

Change-Id: Ic7cc5cec8a26d89a7cea49159ca60a7728e6769c
Bug-Url: https://bugzilla.redhat.com/1627997
Signed-off-by: Pavel Bar <pbar@redhat.com>