api-model: add fips mode to cluster
A new enum intoruced to the engine `FipsMode`. This enum will be
attached to the cluster in order to allow handling of hosts in the
clusters. When set to `UNDEFINED`, once there is a running host on that
cluster it will change to the right FIPS mode derived from the host.
All the hosts in the cluster should have either disabled or enabled FIPS
mode. If they are not aligned with the cluster, they should be
non-responsive. On new cluster when not set the default would be
undefined.

Now we can add a cluster or update it by setting the FIPS mode. Such as:
<fips_mode>disabled</fips_mode>
within the cluster type.

Change-Id: I971ecc6b0c24639821d7fd79d77db8b6201f4ec8
Bug-Url: https://bugzilla.redhat.com/1712481
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>