Add Host DiscoverIscsi Action
This patch adds DiscoverIscsi action and deprecates IscsiDiscover action.

The reason why IscsiDiscover action needs to be replaced rather than modified
is complicated, but in a sentence it stems from an architectural constraint in
the SDKS, whic denies the possibility to deprecate output parameters of an Action.

Due to this limitation this patch:

  https://gerrit.ovirt.org/c/ovirt-engine-api-model/+/84198

which has deprecated the output parameter 'iscsiTargets' and added alongside it
the output parameter 'iscsiDetails', has resulted in below bugzilla.

Bug-Url: https://bugzilla.redhat.com/1926819
Signed-off-by: Ori Liel <oliel@redhat.com>
Change-Id: I02ac1aef9e7d3b3681cafcb4e0f25b6a670663e6