Add BACKUP_SCRATCH disk type to content type
When a backup is taken for a VM, for each disk that participates
in the backup a scratch disk is created.

This patch adds this disk type to the content type so it will be easy
to distinguish between the scratch that was created during a backup
and regular data disks.

Change-Id: I5b12d6d28491c2c0498c4a0ebcadba25f9ab6892
Bug-Url: https://bugzilla.redhat.com/1953558
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>