BackupPhase: add new backup finished phases
After backup finishes, all the execution data disappears.
That means, that the user doesn't know the final execution
state of the operation that was visible via DB and API
while the backup execution was in progress.
The solution is to:
1) Add 2 new phases to show possible execution statuses.
2) Cancel DB cleanup when the backup finishes to allow
DB & API status retrieval by user.
3) Scheduled execution of the cleanup.

This patch adds 2 new phases at the model side:
- SUCCEEDED
- FAILED

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1849861
Change-Id: If0bdca24999fe3d8ac251365d343c5e7e5942e0c