Backup: add a new "modification" date field
Add a new "modification" date field as it was added
to the "vm_backups" DB table.
The field was already added and merged, but in one of the
code reviews it was suggested to change the name of the
column in the database and the related engine code.
Thus, before 4.4.8 is released, it is better to be
consistent and change the name at "api-model" side as well.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1983636
Change-Id: I346caa6ca0b73abe2b502cc4a972531adfc71006