Get Snapshots - include_template flag
When getting the disk snapshots, either for all disks:

  GET .../api/disks/xxx/disksnapshots

of for a storage-domain:

  GET .../api/storagedomains/xxx/disksnapshots

The snapshots of template-disks are not returned.
This poses a problem for code that wishes to
construct the snapshot chain.

A new 'include_template' flag is introcuded. When
set to 'true', template-disk snapshots are included
in the response

Bug-Url: https://bugzilla.redhat.com/1900597
Signed-off-by: Ori Liel <oliel@redhat.com>
Change-Id: I55f796d2c1fed3719c72d1045303672cfd6ebdfd