Add missing DisplayType values
In org.ovirt.engine.core.common.businessentities.DisplayType we have
the supported DisplayType's. The API Model DisplayType values do not
match these values.

The ovirt-engine code contains code to map VNC and SPICE DisplayType
from the API to the businessentities.DisplayType values, but never were
the real values added to the API Model.

In this commit we deprecate the 'legacy' DisplayType's in favor of the
real DisplayType's.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>
Change-Id: Ibc9d7cbec88effd62172983da6ab705160ad8879